cli_bridge: isolate tail index cache (no FastAPI deps for TUI bundle)
