+++ tui_skeleton/docs/subagents_scenarios/ex2_focus_tail_index_cache_capture_20260213.json
@@ -1,6 +1,6 @@
 {
   "scenario": "ex2_focus_tail_index_cache_benchmark",
-  "generatedAt": "2026-02-13T19:51:21Z",
+  "generatedAt": "2026-02-13T20:05:13Z",
   "acceptanceChecks": {
     "returnsTailLines": true,
     "cacheAvoidsRepeatedIndexOps": true,
@@ -18,14 +18,14 @@
       "samples": 40
     },
     "uncached": {
-      "p95Ms": 1.4823899837210774,
-      "maxMs": 1.5529380179941654,
+      "p95Ms": 1.4305210206657648,
+      "maxMs": 1.5228131087496877,
       "indexReadOpsTotal": 40,
       "indexReadBytesTotal": 2621440
     },
     "cached": {
-      "p95Ms": 0.030737021006643772,
-      "maxMs": 0.03886898048222065,
+      "p95Ms": 0.027342932298779488,
+      "maxMs": 0.030957977287471294,
       "indexReadOpsDelta": 0,
       "indexReadBytesDelta": 0
     }
